# [WIP] Compact MetricName tag map implementation