--- a/changelog/@unreleased/pr-1368.v2.yml
+++ b/changelog/@unreleased/pr-1368.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Compact MetricName tag map implementation
+  links:
+  - https://github.com/palantir/tritium/pull/1368