--- a/tritium-jmh/src/jmh/java/com/palantir/tritium/microbenchmarks/NestedMetricsBenchmark.java
+++ b/tritium-jmh/src/jmh/java/com/palantir/tritium/microbenchmarks/NestedMetricsBenchmark.java
@@ -16,11 +16,13 @@
 
 package com.palantir.tritium.microbenchmarks;
 
-import com.codahale.metrics.Meter;
+import com.codahale.metrics.Metric;
 import com.palantir.tritium.metrics.registry.DefaultTaggedMetricRegistry;
 import com.palantir.tritium.metrics.registry.MetricName;
 import com.palantir.tritium.metrics.registry.TaggedMetricRegistry;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
+import java.util.function.BiConsumer;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -38,7 +40,7 @@ import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
 
 @BenchmarkMode(Mode.AverageTime)
-@OutputTimeUnit(TimeUnit.NANOSECONDS)
+@OutputTimeUnit(TimeUnit.MICROSECONDS)
 @Warmup(iterations = 5, time = 1, timeUnit = TimeUnit.SECONDS)
 @Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
 @Fork(1)
@@ -46,26 +48,31 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
 @SuppressWarnings({"designforextension", "NullAway"})
 public class NestedMetricsBenchmark {
     private TaggedMetricRegistry metrics;
+    private BiConsumer<MetricName, Metric> consumer;
 
     @Setup
-    public void before() {
+    public void before(Blackhole blackhole) {
         metrics = constructBaseRegistry();
+        consumer = (name, metric) -> {
+            blackhole.consume(name.safeName());
+            // name.safeTags().forEach(blackholeConsumer) would be ideal,
+            // however most readers will use the entrySet
+            for (Map.Entry<String, String> entry : name.safeTags().entrySet()) {
+                blackhole.consume(entry.getKey());
+                blackhole.consume(entry.getValue());
+            }
+            blackhole.consume(metric);
+        };
     }
 
     @Benchmark
-    public void benchmarkGetMetrics(Blackhole blackhole) {
-        metrics.getMetrics().forEach((name, metric) -> {
-            blackhole.consume(name);
-            blackhole.consume(metric);
-        });
+    public void benchmarkGetMetrics() {
+        metrics.getMetrics().forEach(consumer);
     }
 
     @Benchmark
-    public void benchmarkForEach(Blackhole blackhole) {
-        metrics.forEachMetric((name, metric) -> {
-            blackhole.consume(name);
-            blackhole.consume(metric);
-        });
+    public void benchmarkForEach() {
+        metrics.forEachMetric(consumer);
     }
 
     private static TaggedMetricRegistry constructBaseRegistry() {
@@ -79,13 +86,13 @@ public class NestedMetricsBenchmark {
     private static TaggedMetricRegistry constructSubRegistry() {
         TaggedMetricRegistry registry = new DefaultTaggedMetricRegistry();
         for (int i = 0; i < 100; i++) {
-            Meter meter = registry.meter(MetricName.builder()
-                    .safeName("some metric " + i)
-                    .putSafeTags("some tag", "some tag value")
-                    .build());
-            for (int j = 0; j < 1000; j++) {
-                meter.mark();
-            }
+            registry.meter(MetricName.builder()
+                            .safeName("some metric " + i)
+                            .putSafeTags("some tag", "some tag value")
+                            .putSafeTags("libraryName", "tritium")
+                            .putSafeTags("libraryVersion", "1.2.3")
+                            .build())
+                    .mark();
         }
         return registry;
     }