--- a/tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/ExtraEntrySortedMap.java
+++ b/tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/ExtraEntrySortedMap.java
@@ -20,7 +20,6 @@ import static com.palantir.logsafe.Preconditions.checkArgument;
 import static com.palantir.logsafe.Preconditions.checkNotNull;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Ordering;
@@ -33,6 +32,7 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 @SuppressWarnings("JdkObsolete")
@@ -58,6 +58,7 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
         return base.comparator();
     }
 
+    @Nonnull
     @Override
     public SortedMap<K, V> subMap(K fromKey, K toKey) {
         SortedMap<K, V> newBase = base.subMap(fromKey, toKey);
@@ -67,6 +68,7 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
         return newBase;
     }
 
+    @Nonnull
     @Override
     public SortedMap<K, V> headMap(K toKey) {
         SortedMap<K, V> newBase = base.headMap(toKey);
@@ -76,6 +78,7 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
         return newBase;
     }
 
+    @Nonnull
     @Override
     public SortedMap<K, V> tailMap(K fromKey) {
         SortedMap<K, V> newBase = base.tailMap(fromKey);
@@ -141,7 +144,7 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
     }
 
     @Override
-    public void putAll(Map<? extends K, ? extends V> _other) {
+    public void putAll(@Nonnull Map<? extends K, ? extends V> _other) {
         throw new UnsupportedOperationException();
     }
 
@@ -150,13 +153,17 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
         throw new UnsupportedOperationException();
     }
 
+    @Nonnull
     @Override
     public Set<K> keySet() {
-        return new AbstractSet<K>() {
+        return new AbstractSet<>() {
             @Override
             public Iterator<K> iterator() {
-                return Iterables.mergeSorted(ImmutableList.of(base.keySet(), ImmutableList.of(extraKey)), ordering)
-                        .iterator();
+                return Iterators.mergeSorted(
+                        ImmutableList.of(
+                                base.keySet().iterator(),
+                                ImmutableList.of(extraKey).iterator()),
+                        ordering);
             }
 
             @Override
@@ -166,9 +173,10 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
         };
     }
 
+    @Nonnull
     @Override
     public Collection<V> values() {
-        return new AbstractCollection<V>() {
+        return new AbstractCollection<>() {
             @Override
             public Iterator<V> iterator() {
                 return Iterators.transform(keySet().iterator(), key -> get(key));
@@ -181,9 +189,10 @@ final class ExtraEntrySortedMap<K, V> extends AbstractMap<K, V> implements Sorte
         };
     }
 
+    @Nonnull
     @Override
     public Set<Entry<K, V>> entrySet() {
-        return new AbstractSet<Map.Entry<K, V>>() {
+        return new AbstractSet<>() {
             @Override
             public Iterator<Map.Entry<K, V>> iterator() {
                 return Iterators.transform(keySet().iterator(), key -> Maps.immutableEntry(key, get(key)));