… tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/PrehashedSortedMap.java
… tritium-registry/src/test/java/com/palantir/tritium/metrics/registry/ExtraEntrySortedMapTest.java
… tritium-registry/src/test/java/com/palantir/tritium/metrics/registry/TagMapTest.java