## Before this PR
Excessive cost to create and store MetricName due to tags.

## After this PR
==COMMIT_MSG==
Compact MetricName tag map implementation
==COMMIT_MSG==

## Possible downsides?
Custom code