@carterkozak 2022-03-01
I'm preparing a change to remove the ExtraEntryMap wrapper in favor of using TagMap directly