@carterkozak 2022-03-01
# NestedMetricsBenchmark results

Benchmarks after:
```
Benchmark                                                                    Mode  Cnt        Score        Error   Units
NestedMetricsBenchmark.benchmarkForEach                                      avgt   10   292477.731 ±  12461.604   ns/op
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate.norm                  avgt   10   362048.602 ±      4.604    B/op
NestedMetricsBenchmark.benchmarkGetMetrics                                   avgt   10   610141.258 ±   3238.038   ns/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate.norm               avgt   10  1040735.489 ±     15.408    B/op
```

Benchmarks before:
```
Benchmark                                                                    Mode  Cnt        Score        Error   Units
NestedMetricsBenchmark.benchmarkForEach                                      avgt   10   295342.347 ±   6560.391   ns/op
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate.norm                  avgt   10   324494.444 ±      8.378    B/op
NestedMetricsBenchmark.benchmarkGetMetrics                                   avgt   10   623551.299 ±  25264.115   ns/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate.norm               avgt   10  1000733.243 ±     18.299    B/op
```

## Analysis

These don't seem to tell the whole story, they measure object creation, but we never consume the metricname components, there's no iteration over keys.

## More representative benchmarks

With the benchmark changes introduced in 173efe9 we have larger tag maps, and the consumer blackholes the safeName and all tag key/value pairs.

## After:

### tags forEach(biconsumer)

```
Benchmark                                                                    Mode  Cnt        Score        Error   Units
NestedMetricsBenchmark.benchmarkForEach                                      avgt   10      398.119 ±      2.923   us/op
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate                       avgt   10      513.892 ±      3.774  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate.norm                  avgt   10   322031.270 ±      8.948    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space              avgt   10      512.321 ±    287.807  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space.norm         avgt   10   320913.039 ± 179534.072    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space          avgt   10        0.009 ±      0.025  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space.norm     avgt   10        5.628 ±     15.830    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.count                            avgt   10       13.000               counts
NestedMetricsBenchmark.benchmarkForEach:·gc.time                             avgt   10       18.000                   ms
NestedMetricsBenchmark.benchmarkGetMetrics                                   avgt   10      777.042 ±     13.052   us/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate                    avgt   10      916.455 ±     15.157  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate.norm               avgt   10  1120738.250 ±     16.985    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space           avgt   10      906.663 ±    287.963  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space.norm      avgt   10  1108490.138 ± 348872.370    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space       avgt   10        0.234 ±      0.089  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space.norm  avgt   10      286.871 ±    110.637    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.count                         avgt   10       23.000               counts
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.time                          avgt   10       31.000                   ms
```

### tags entrySet + enhanced for-loop
```
Benchmark                                                                    Mode  Cnt        Score        Error   Units
NestedMetricsBenchmark.benchmarkForEach                                      avgt   10      404.491 ±     15.991   us/op
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate                       avgt   10      506.065 ±     19.342  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate.norm                  avgt   10   322031.456 ±      8.778    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space              avgt   10      512.328 ±    287.825  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space.norm         avgt   10   325076.827 ± 176804.791    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space          avgt   10        0.009 ±      0.024  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space.norm     avgt   10        5.472 ±     15.307    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.count                            avgt   10       13.000               counts
NestedMetricsBenchmark.benchmarkForEach:·gc.time                             avgt   10       19.000                   ms
NestedMetricsBenchmark.benchmarkGetMetrics                                   avgt   10      809.683 ±     24.776   us/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate                    avgt   10      879.724 ±     26.745  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate.norm               avgt   10  1120738.161 ±     16.376    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space           avgt   10      866.973 ±    251.475  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space.norm      avgt   10  1104453.933 ± 317276.137    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space       avgt   10        0.200 ±      0.117  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space.norm  avgt   10      255.357 ±    152.886    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.count                         avgt   10       22.000               counts
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.time                          avgt   10       29.000                   ms
```
## Before:

### tags forEach(biconsumer)
```
Benchmark                                                                    Mode  Cnt        Score        Error   Units
NestedMetricsBenchmark.benchmarkForEach                                      avgt   10     1496.868 ±     44.749   us/op
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate                       avgt   10     1241.821 ±     36.396  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate.norm                  avgt   10  2924587.239 ±     19.219    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space              avgt   10     1221.408 ±    188.776  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space.norm         avgt   10  2877447.293 ± 451164.940    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space          avgt   10        0.007 ±      0.015  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space.norm     avgt   10       16.604 ±     33.975    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.count                            avgt   10       31.000               counts
NestedMetricsBenchmark.benchmarkForEach:·gc.time                             avgt   10       37.000                   ms
NestedMetricsBenchmark.benchmarkGetMetrics                                   avgt   10     1864.543 ±     48.843   us/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate                    avgt   10     1227.344 ±     31.588  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate.norm               avgt   10  3600849.967 ±     28.255    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space           avgt   10     1221.425 ±    188.120  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space.norm      avgt   10  3582756.816 ± 524419.814    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space       avgt   10        0.039 ±      0.038  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space.norm  avgt   10      115.083 ±    113.371    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.count                         avgt   10       31.000               counts
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.time                          avgt   10       41.000                   ms
```

### tags entrySet + enhanced for-loop

```
Benchmark                                                                    Mode  Cnt        Score        Error   Units
NestedMetricsBenchmark.benchmarkForEach                                      avgt   10     1610.645 ±     69.380   us/op
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate                       avgt   10     1202.110 ±     51.561  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.alloc.rate.norm                  avgt   10  3044592.417 ±      5.200    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space              avgt   10     1182.330 ±      0.580  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Eden_Space.norm         avgt   10  2996674.988 ± 129049.113    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space          avgt   10        0.008 ±      0.019  MB/sec
NestedMetricsBenchmark.benchmarkForEach:·gc.churn.G1_Survivor_Space.norm     avgt   10       20.970 ±     49.960    B/op
NestedMetricsBenchmark.benchmarkForEach:·gc.count                            avgt   10       30.000               counts
NestedMetricsBenchmark.benchmarkForEach:·gc.time                             avgt   10       34.000                   ms
NestedMetricsBenchmark.benchmarkGetMetrics                                   avgt   10     1962.866 ±    102.056   us/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate                    avgt   10     1205.972 ±     60.899  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.alloc.rate.norm               avgt   10  3720853.849 ±     16.751    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space           avgt   10     1181.870 ±      0.728  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Eden_Space.norm      avgt   10  3650231.367 ± 188346.090    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space       avgt   10        0.146 ±      0.189  MB/sec
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.churn.G1_Survivor_Space.norm  avgt   10      453.699 ±    590.895    B/op
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.count                         avgt   10       30.000               counts
NestedMetricsBenchmark.benchmarkGetMetrics:·gc.time                          avgt   10       38.000                   ms
```